+++ src/components/IndexPage.components.tsx
@@ -8,9 +8,9 @@ interface ReactComponentWithLang {
 
 export const ProfileImage: React.FC<ReactComponentWithLang> = ({ lang }) => {
     return (
-        <div className='justify-center items-center bg-gradient-to-br from-[#f28e26] to-[#68d8d6] w-[8.2rem] h-[8.2rem] tablet:w-[13.2rem] tablet:h-[13.2rem] rounded-full mt-6 mb-4'>
-            <div className='justify-center items-center bg-main-white w-[7.8rem] h-[7.8rem] tablet:w-[12.8rem] tablet:h-[12.8rem] rounded-full'>
-                <img src={ProfileImageFile} alt={getLocaleString('profileimage_alt', lang)} className='w-[7.5rem] tablet:w-[12rem] rounded-full' />
+        <div className='justify-center items-center bg-gradient-to-br from-[#f28e26] to-[#68d8d6] w-[8.2rem] h-[8.2rem] tablet:w-[10.2rem] tablet:h-[10.2rem] rounded-full mt-6 mb-4'>
+            <div className='justify-center items-center bg-main-white w-[7.8rem] h-[7.8rem] tablet:w-[9.8rem] tablet:h-[9.8rem] rounded-full'>
+                <img src={ProfileImageFile} alt={getLocaleString('profileimage_alt', lang)} className='w-[7.5rem] tablet:w-[9rem] rounded-full' />
             </div>
         </div>
     )
@@ -67,7 +67,7 @@ export const MyOther: React.FC<ReactComponentWithLang> = ({ lang }) => {
         <div className="items-center w-full mb-4">
             <h1 className="font-bold text-xs text-main-black mb-2">{getLocaleString('my_other', lang, true)}</h1>
             <LinkButton type="gmail" />
-            <LinkButton type="ngl" />
+            {/* <LinkButton type="ngl" /> */}
         </div>
     )
 }