+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
-## December 18th, 2022
+## December 20th, 2022: *(0.1.1)*
+- Better components structure.
+- "Copy User to Clipboard" functionality now works correctly.
+
+## December 18th, 2022: *(0.1.0)*
 - Initial commit
     - First environment deployment.